+++ R/conversions.R
@@ -51,6 +51,125 @@ sampled.tree.from.combined = function(tree, rho = 1, sampled_tips = NULL) {
   tree
 }
 
+#' Returns tree and fossil objects that you can use to plot the reconstructed tree.
+#'
+#' Note that for datasets containing extinct only samples (& rho = 0) the ages output are scaled so that the youngest sample = 0.
+#'
+#' @param tree Tree object.
+#' @param fossils Fossils object.
+#' @param rho Extant species sampling probability. Default = 1, will be disregarded if fossils object already contains extant samples.
+#'
+#' @return A list containing the tree and fossil objects.
+#'
+#' @examples
+#' # simulate tree
+#' birth = 0.1
+#' death = 0.05
+#' tips = 10
+#' t = TreeSim::sim.bd.taxa(tips, 1, birth, death)[[1]]
+#'
+#' # simulate fossils
+#' f = sim.fossils.poisson(rate = 0.3, tree = t)
+#'
+#' # simulate extant samples
+#' f = sim.extant.samples(f, tree = t, rho = 0.5)
+#'
+#' # plot the complete tree
+#' plot(f,t)
+#'
+#' # generate tree & fossil objects corresponding to the reconstructed tree
+#' out = reconstructed.tree.fossils.objects(f, t)
+#' f.reconst = out$fossils
+#' t.reconst = out$tree
+#'
+#' # plot the reconstructed tree
+#' plot(f.reconst, t.reconst)
+#'
+#' @export
+# function to generate tree and corresponding fossil object for the reconstructed tree
+reconstructed.tree.fossils.objects = function(fossils, tree, rho = 1){
+
+  if(!is.null(tree) && !"phylo" %in% class(tree))
+    stop("tree must be an object of class \"phylo\"")
+
+  if(!is.null(fossils) && !"fossils" %in% class(fossils))
+    stop("fossils must be an object of class \"fossils\"")
+
+  if(!(rho >= 0 && rho <= 1))
+    stop("rho must be a probability between 0 and 1")
+
+  tol = min((min(tree$edge.length)/100),1e-8)
+  samp_tips = NULL
+
+  if(any( abs(fossils$hmax) < tol )){
+    # identify extant sampled tips
+    samp_tips = tree$tip.label[fossils$edge[which(abs(fossils$hmax) < tol)]]
+    # eliminate them from the dataframe (required for SAtree.from.fossils)
+    fossils = fossils[-c(which(abs(fossils$hmax) < tol)),]
+  }
+
+  # create SAtree object
+  sa.tree = SAtree.from.fossils(tree, fossils)
+
+  # match samp_tips and sa.tree tip labels
+  if(!is.null(samp_tips)){
+    for(i in 1:length(samp_tips)){
+      samp_tips[i] = sa.tree$tip.label [ grep(paste0(samp_tips[i],"_"), sa.tree$tip.label)[1] ]
+    }
+  }
+
+  # removes all unsampled lineages
+  samp.tree = sampled.tree.from.combined(sa.tree, rho = rho, sampled_tips = samp_tips)
+
+  # identify sampled ancestors and drop from the sampled tree
+  sa = c()
+  for(i in 1:length(samp.tree$tip.label)){
+    blength = samp.tree$edge.length[which(samp.tree$edge[,2]==i)]
+    if(abs(blength) < tol) sa = c(sa, samp.tree$tip.label[i])
+  }
+  no.sa.tree = ape::drop.tip(samp.tree, sa)
+
+  # create new fossils object based on the reconstructed tree
+  # & deal with sampled ancestors
+  if(length(fossils$sp) > 0){
+    f.new = data.frame()
+    nages = n.ages(samp.tree)
+    for(i in sa){
+      anc = ancestor(which(samp.tree$tip.label==i), samp.tree)
+
+      s1 = ape::extract.clade(samp.tree,anc)$tip.label
+      s2 = no.sa.tree$tip.label[which(no.sa.tree$tip.label %in% s1)]
+
+      # assign fossils to nearest descendent node in the tree
+      if(length(s2) > 1){
+        j = ape::getMRCA(no.sa.tree,s2)
+      } else { # i is SA on terminal branch
+        j = which(no.sa.tree$tip.label==s2)
+      }
+      h = nages[[which(samp.tree$tip.label==i)]]
+      f.new = rbind(f.new, data.frame(sp = j,
+                                      edge = j,
+                                      hmin = h,
+                                      hmax = h))
+    }
+    f.new = fossils(f.new)
+  } else {
+    f.new = fossils()
+  }
+
+  # deal with non SA samples
+  f.new = FossilSim::sim.tip.samples(f.new, no.sa.tree)
+
+  # add a root edge for any stem fossils
+  if( any(f.new$edge == (length(no.sa.tree$tip.label)+1) )){
+    h = f.new$hmin[which(f.new$edge == (length(no.sa.tree$tip.label)+1))]
+    no.sa.tree$root.edge = max(h) - max(n.ages(no.sa.tree))
+  } else { no.sa.tree$root.edge = NULL }
+
+  return(list(tree = no.sa.tree, fossils = f.new))
+
+}
+
 #' Removes all intermediate fossils from a combined tree and labels the first and last fossils of each lineage.
 #' Can be used with sampled or complete trees. If only one fossil is present for a particular species it is labelled as first.
 #'
@@ -115,7 +234,9 @@ prune.fossils = function(tree) {
 #'
 #' # output for BEAST
 #' beast.fbd.format(t, f) # output on the console
+#' \dontrun{
 #' beast.fbd.format(t, f, file="example.tre") # output in file
+#' }
 #' @export
 beast.fbd.format = function(tree, fossils, rho = 1, sampled_tips = NULL, ...) {
   proc_tree = prune.fossils(sampled.tree.from.combined(SAtree.from.fossils(tree,fossils), rho = rho, sampled_tips = sampled_tips))
@@ -388,6 +509,7 @@ fossils.to.paleotree.record = function(fossils, tree = NULL, taxonomy = NULL) {
 #' @param min Value used to represent the minimum possible interval age of extinct specimens with \code{hmin = 0}. By default \code{min = NULL} and the function will use the sampling times in the fossils dataframe.
 #' @param exclude.extant.singletons If TRUE exclude species that have extant samples only (default = TRUE).
 #' @param file Output file name.
+#' @param use.sp.names If TRUE use the value in fossils$sp as the complete taxon name, otherwise the function adds the prefix "taxa" (default = FALSE).
 #'
 #' @examples
 #'
@@ -423,7 +545,7 @@ fossils.to.paleotree.record = function(fossils, tree = NULL, taxonomy = NULL) {
 #'
 #' @export
 fossils.to.pyrate = function(fossils, python = TRUE, traits = NULL, cutoff = NULL, random = FALSE, min = NULL,
-                             exclude.extant.singletons = TRUE, file = ""){
+                             exclude.extant.singletons = TRUE, file = "", use.sp.names = FALSE){
 
   if(length(fossils$sp) < 1) stop("Number of specimens must be > 0")
 
@@ -459,7 +581,8 @@ fossils.to.pyrate = function(fossils, python = TRUE, traits = NULL, cutoff = NUL
       fossils$hmin = round(unlist(lapply(1:dim(fossils)[1], function(x) { mean( c(fossils$hmin[x], fossils$hmax[x])) })), digits = 6)
       fossils$hmax = fossils$hmin
     } else {
-      fossils = round(fossils, digits = 6)
+      fossils$hmin = round(fossils$hmin, digits = 6)
+      fossils$hmax = round(fossils$hmax, digits = 6)
     }
   }
 
@@ -487,7 +610,10 @@ fossils.to.pyrate = function(fossils, python = TRUE, traits = NULL, cutoff = NUL
 
       if(length(times) > 0){
         data_1 = c(data_1, paste0("array([", paste(times, collapse = ","), "])"))
-        taxa_names = c(taxa_names, paste0("'taxa", sp, "'"))
+        if(use.sp.names)
+          taxa_names = c(taxa_names, paste0("'", sp, "'"))
+        else
+          taxa_names = c(taxa_names, paste0("'taxa", sp, "'"))
         if(!is.null(traits)) trait1 = c(trait1, paste0("array([", paste(values, collapse = ","), "])"))
       }
     }
@@ -495,7 +621,7 @@ fossils.to.pyrate = function(fossils, python = TRUE, traits = NULL, cutoff = NUL
     # print fossil ages
     cat(paste(data_1, collapse = ",\n"), "]\n", sep = "\n", file = file, append = TRUE)
 
-    cat("d=[data_1]", "names=['example_1']", "def get_data(i): return d[i]", "def get_out_name(i): return  names[i]",
+    cat("d=[data_1]", paste0("names=['", file ,"']"), "def get_data(i): return d[i]", "def get_out_name(i): return  names[i]",
         sep = "\n", file = file, append = TRUE)
 
     # print species names